fix: revert ApiInfo bean move to SwaggerAutoConfiguration
