+++ activiti-cloud-query-service/activiti-cloud-starter-query/src/main/java/org/activiti/cloud/starter/query/configuration/QuerySwaggerConfig.java
@@ -18,13 +18,30 @@ package org.activiti.cloud.starter.query.configuration;
 
 import java.util.function.Predicate;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.info.BuildProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import springfox.documentation.RequestHandler;
+import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
 
 @Configuration
 public class QuerySwaggerConfig {
+
+    @Bean
+    public ApiInfo apiInfo(BuildProperties buildProperties) {
+        return new ApiInfoBuilder()
+            .title(String.format("%s ReST API", buildProperties.getName()))
+            .description(buildProperties.get("description"))
+            .version(buildProperties.getVersion())
+            .license(String.format("© %s-%s %s. All rights reserved",
+                buildProperties.get("inceptionYear"),
+                buildProperties.get("year"),
+                buildProperties.get("organization.name")))
+            .termsOfServiceUrl(buildProperties.get("organization.url"))
+            .build();
+    }
 
     @Bean
     @ConditionalOnMissingBean
